requirements.txt
# JaalTaka - Counterfeit Banknote Detection
# Core ML
torch>=2.1.0
torchvision>=0.16.0
timm>=0.9.12

# Data & Science
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Explainability
grad-cam>=1.5.0
lime>=0.2.0.1
shap>=0.43.0
captum>=0.7.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Export & Compression
onnx>=1.15.0
onnxruntime>=1.16.0
# tf-nightly or tensorflow for TFLite conversion
# tensorflow>=2.15.0

# Experiment Tracking
tensorboard>=2.15.0
tqdm>=4.66.0

# Utilities
PyYAML>=6.0
tabulate>=0.9.0
